backend/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
scikit-image
reportlab
python-multipart
